파일 업로드 작업 백업

+++ src/main/java/com/practice/board/application/board/dao/BoardAttachMapper.java
@@ -31,5 +31,5 @@ public interface BoardAttachMapper {
      * @param boardIdx
      * @return List<BoardAttach>
      */
-    List<BoardAttach> findByIdx(Long boardIdx);
+    List<BoardAttach> findByIdx(int boardIdx);
 }
